fix:完善social req和resp

+++ biz/dao/social.go
@@ -15,21 +15,21 @@ func (m *MySQLdb) DeleteFollowing(userId string, toUserId string) error {
 	_, err := m.db.Exec(sql, userId, toUserId)
 	return err
 }
-func (m *MySQLdb) FollowingList(userId string, pageNum int, pageSize int) ([]entity.UserEntity, error) {
+func (m *MySQLdb) FollowingList(userId string, pageNum int64, pageSize int64) ([]entity.UserEntity, error) {
 	sql := `SELECT * FROM users WHERE id IN (SELECT following_id FROM following WHERE user_id = ?) LIMIT ? OFFSET ?`
 	var users []entity.UserEntity
 	offset := pageNum * pageSize
 	err := m.db.Select(&users, sql, userId, pageSize, offset)
 	return users, err
 }
-func (m *MySQLdb) FollowerList(userId string, pageNum int, pageSize int) ([]entity.UserEntity, error) {
+func (m *MySQLdb) FollowerList(userId string, pageNum int64, pageSize int64) ([]entity.UserEntity, error) {
 	sql := `SELECT * FROM users WHERE id IN (SELECT user_id FROM following WHERE following_id = ?) LIMIT ? OFFSET ?`
 	offset := pageNum * pageSize
 	var users []entity.UserEntity
 	err := m.db.Select(&users, sql, userId, pageSize, offset)
 	return users, err
 }
-func (m *MySQLdb) FriendList(userId string, pageNum int, pageSize int) ([]entity.UserEntity, bool) {
+func (m *MySQLdb) FriendList(userId string, pageNum int64, pageSize int64) ([]entity.UserEntity, bool) {
 	offset := pageNum * pageSize
 	sql := `SELECT * FROM users WHERE id IN(SELECT following_id FROM following WHERE user_id = ?) 
                       AND id IN (SELECT user_id from following where following_id=?) LIMIT ? OFFSET ?`
